superfluous information without explanation
the comparison to POST is distracting without enumerating differences (if any), which would probably be more distracting in this context as much depends on the the particular HTTP server's expectations. i feel this paragraph's cogency can be improved by shortening it thus. perhaps someday i or someone else can hyperlink to more information on the differences.