superfluous distinction made between GET and POST
this paragraph can be improved by 1) omitting the declaration that GET is not POST (since the only difference may be the method text itself, as servers decide what to do with the payload); and 2) by clarifying that the request is sent "AS IF" from a web browser/client software that exposes the form data locally in the url. it is slightly misleading to say the data is appended to the right of the URL unless there is some way curl can reference the URL in the same command or say cache it. and it might be better to indicate it is added the same way to the payload body as with POST. forgive me if suggestion #2 represents an oversight of cur's capabilities on my part and thank you for considering my two change recommendations, as reflected.