superfluous statement and unclear sentence (#137)
* superfluous statement and unclear sentence

this paragraph can be improved by 1) omitting the statement that POST is not GET (this is fundamentally true insofar as the text of the method is one or the other) and 2) clarifying that the convenience comes from reusing the field information from a POST so that the luser doesn't have to write it out at the end of the url on the command line. i have proposed my wording thus to accomplish these improvements. thank you for considering my suggestions.

i suggested command-line URL when it is probably more relevant to say inputted URL assuming this information applies to libcurl options.