tweak table headers
Because of some peculiarity in pandoc it easily generates table columns
that overlap when generating the PDF, which apparently can be tweaked by
editing the length of the dashes on the top of the table.

But also: these tables possibly also show that they are not nice to do
as tables when getting displayed in a page in a PDF as they have
contents that make the columns very narrow. We should reconsider turning
them into lists instead.

Fixes #418