SUMMARY: major reorg
I want to make the different sections and what is put where more
obvious.

The command line sections are now:

 - Command line concepts
 - Command line transfers
 - Command line HTTP
 - Command line FTP

Named like this to stress command line. The first describes concepts,
the second how to run and control transfers in general while the two
latter ones speak of specific options and things unique for HTTP and
FTP.

The libcurl sections gets a similar shakedown, and the new layout works
as per below:

 - libcurl: for top-level description of library concepts
 - libcurl transfers: use the API to drive and control transfers
 - libcurl HTTP: details specific to HTP transfers
 - libcurl helpers: non-transfer focused APIs described (URL API, share
   API, WebSocket etc)
 - libcurl examples: stand-alone small applications showing off libcurl use
 - libcurl bindings: the bindings
 - libcurl internals: the internals

Additionally:

All subdirectories now has a README.md which is the leading text for
search sub-section. This moves the first text into their own subdirs,
for easier management and this also allows for better/easier browsing
the content directly off filesystems and directly on GitHub.