no longer link to README.md
With the transition to mdbook, links to ../README.md no longer work, as
they are now converted to 'index.html' in the web version. This change
now instead links to "dir/" without a file name in an attempt to make
the markdown version on github also have working links.

The downside is that a locally rendered book that you view on file://
gets broken links.

Closes #441